Move behind experimentalOllamaChat config
