+++ lib/shared/src/models/index.ts
@@ -1,4 +1,5 @@
-import { OLLAMA_DEFAULT_URL } from '../ollama/completions-client'
+import { logError } from '../logger'
+import { OLLAMA_DEFAULT_URL } from '../ollama'
 import { isDotCom } from '../sourcegraph-api/environments'
 import { DEFAULT_DOT_COM_MODELS } from './dotcom'
 import { ModelUsage } from './types'
@@ -31,31 +32,47 @@ export class ModelProvider {
     // Providers available for dotcom instances
     private static dotComProviders: ModelProvider[] = DEFAULT_DOT_COM_MODELS
     // Providers available from local ollama instances
+    private static ollamaProvidersEnabled = false
     private static ollamaProviders: ModelProvider[] = []
+
+    public static onConfigChange(enableOllamaModels: boolean): void {
+        ModelProvider.ollamaProvidersEnabled = enableOllamaModels
+        ModelProvider.ollamaProviders = []
+        if (enableOllamaModels) {
+            ModelProvider.getLocalOllamaModels()
+        }
+    }
 
     /**
      * Fetches available Ollama models from the local Ollama server
      * and adds them to the list of ollama providers.
      */
-    public static getLocalOllamaModels(endpoint?: string | null): void {
+    public static getLocalOllamaModels(): void {
         const isAgentTesting = process.env.CODY_SHIM_TESTING === 'true'
-        // Only fetch local models if the endpoint is not a dotcom instance
-        if (isAgentTesting || !endpoint || !isDotCom(endpoint)) {
+        // Only fetch local models if user has enabled the config
+        if (isAgentTesting || !ModelProvider.ollamaProvidersEnabled) {
             return
         }
-
+        // TODO (bee) watch file change to determine if a new model is added
+        // to eliminate the needs of restarting the extension to get the new models
         fetch(new URL('/api/tags', OLLAMA_DEFAULT_URL).href)
             .then(response => response.json())
-            .then(data => {
-                const models = new Set<ModelProvider>()
-                for (const model of data.models) {
-                    const name = `ollama/${model.model}`
-                    const newModel = new ModelProvider(name, [ModelUsage.Chat])
-                    models.add(newModel)
+            .then(
+                data => {
+                    const models = new Set<ModelProvider>()
+                    for (const model of data.models) {
+                        const name = `ollama/${model.model}`
+                        const newModel = new ModelProvider(name, [ModelUsage.Chat, ModelUsage.Edit])
+                        models.add(newModel)
+                    }
+                    ModelProvider.ollamaProviders = Array.from(models)
+                },
+                error => {
+                    const isFetchFailed = error.toString().endsWith('fetch failed')
+                    const serverErrorMsg = 'Please make sure the Ollama server is up & running.'
+                    logError('getLocalOllamaModels: failed ', isFetchFailed ? serverErrorMsg : error)
                 }
-                ModelProvider.ollamaProviders = Array.from(models)
-            })
-            .catch(() => console.log('Cannot find local ollama models'))
+            )
     }
 
     /**
@@ -84,9 +101,11 @@ export class ModelProvider {
         const isDotComUser = !endpoint || (endpoint && isDotCom(endpoint))
         const models = (
             isDotComUser
-                ? ModelProvider.dotComProviders.concat(ModelProvider.ollamaProviders)
+                ? ModelProvider.dotComProviders
                 : Array.from(ModelProvider.privateProviders.values())
-        ).filter(model => model.usage.includes(type))
+        )
+            .concat(ModelProvider.ollamaProviders)
+            .filter(model => model.usage.includes(type))
 
         if (!isDotComUser) {
             return models
