+++ lib/shared/src/ollama/completions-client.ts
@@ -1,5 +1,6 @@
+import type { OllamaGenerateErrorResponse, OllamaGenerateParams, OllamaGenerateResponse } from '.'
 import { isDefined } from '../common'
-import type { OllamaGenerateParameters, OllamaOptions } from '../configuration'
+import type { OllamaOptions } from '../configuration'
 import {
     type CodeCompletionsClient,
     type CompletionResponseGenerator,
@@ -10,65 +11,6 @@ import type { CompletionResponse } from '../sourcegraph-api/completions/types'
 import { isAbortError } from '../sourcegraph-api/errors'
 import { isNodeResponse } from '../sourcegraph-api/graphql/client'
 import { isError } from '../utils'
-
-export const OLLAMA_DEFAULT_URL = 'http://localhost:11434'
-
-/**
- * @see https://sourcegraph.com/github.com/jmorganca/ollama/-/blob/api/types.go?L35
- */
-export interface OllamaGenerateParams {
-    model: string
-    template: string
-    prompt: string
-    options?: OllamaGenerateParameters
-}
-
-/**
- * @see https://sourcegraph.com/github.com/jmorganca/ollama/-/blob/api/types.go?L50
- */
-export interface OllamaChatParams {
-    model: string
-    messages: OllamaChatMessage[]
-    format?: string
-    stream?: boolean
-    //  controls how long the model will stay loaded into memory following the request (default: 5m)
-    keep_alive?: string
-
-    options?: OllamaGenerateParameters
-}
-
-interface OllamaChatMessage {
-    role: string
-    content: string
-    images?: string[]
-}
-
-/**
- * @see https://sourcegraph.com/github.com/jmorganca/ollama/-/blob/api/types.go?L88
- */
-export interface OllamaGenerateResponse {
-    model: string
-    response?: string
-    done: boolean
-    context?: number[]
-    total_duration?: number
-    load_duration?: number
-    prompt_eval_count?: number
-    prompt_eval_duration?: number
-    eval_count?: number
-    eval_duration?: number
-    sample_count?: number
-    sample_duration?: number
-    message?: {
-        role: string
-        content: string
-        images?: string[] | null
-    }
-}
-
-interface OllamaGenerateErrorResponse {
-    error?: string
-}
 
 const RESPONSE_SEPARATOR = /\r?\n/
