+++ lib/shared/src/telemetry/EventLogger.ts
@@ -120,6 +120,7 @@ export class EventLogger {
             configurationDetails: {
                 contextSelection: this.config.useContext,
                 guardrails: this.config.experimentalGuardrails,
+                ollama: this.config.experimentalOllamaChat,
             },
             version: this.extensionDetails.version, // for backcompat
             hasV2Event,
