+++ vscode/src/main.ts
@@ -4,6 +4,7 @@ import {
     type ChatEventSource,
     ConfigFeaturesSingleton,
     type ConfigurationWithAccessToken,
+    ModelProvider,
     PromptMixin,
     featureFlagProvider,
     graphqlClient,
@@ -229,6 +230,9 @@ const register = async (
         if (oldConfig === JSON.stringify(newConfig)) {
             return Promise.resolve()
         }
+
+        ModelProvider.onConfigChange(newConfig.experimentalOllamaChat)
+
         const promises: Promise<void>[] = []
         oldConfig = JSON.stringify(newConfig)
 
@@ -301,6 +305,7 @@ const register = async (
     })
     // Sync initial auth status
     await chatManager.syncAuthStatus(authProvider.getAuthStatus())
+    ModelProvider.onConfigChange(initialConfig.experimentalOllamaChat)
 
     const commandsManager = platform.createCommandsProvider?.()
     setCommandController(commandsManager)
